requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==11.1.0
pynput==1.8.1
python-xlib==0.33
requests==2.32.3
six==1.17.0
urllib3==2.4.0
